Refactoring horizontal scaling adapters
