+++ src/adapters/index.ts
@@ -1,5 +1,6 @@
 export * from './adapter';
 export * from './adapter-interface';
 export * from './cluster-adapter';
+export * from './horizontal-adapter';
 export * from './local-adapter';
 export * from './redis-adapter';
